Split test_download.py
